pyproject.toml
[project]
name = "py-sync-settings"
version = "0.1.0"
description = ""
authors = [
    {name = "serii",email = "seriiburduja@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.10,<4.0"
dependencies = [
    "autopep8 (==2.3.2)",
    "flake8 (==7.2.0)",
    "gitdb (==4.0.11)",
    "gitpython==3.1.43",
    "markdown-it-py (==3.0.0)",
    "mccabe (==0.7.0)",
    "mdurl (==0.1.2)",
    "mypy (==1.16.0)",
    "mypy-extensions (==1.1.0)",
    "pathspec (==0.12.1)",
    "plyer (==2.1.0)",
    "pycodestyle (==2.13.0)",
    "pyflakes (==3.3.2)",
    "pyfzf (==0.3.1)",
    "pygments (==2.18.0)",
    "pyperclip (==1.9.0)",
    "rich (==13.8.1)",
    "smmap (==5.0.1)",
    "typing-extensions (==4.14.0)",
]


